data_collection/Social_media_crawler/requirements.txt
aiofiles==23.2.1
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.110.2
fonttools==4.60.2
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
jieba==0.42.1
jiter==0.12.0
jmespath==1.0.1
kiwisolver==1.4.7
loguru==0.7.3
lxml==6.0.2
matplotlib==3.9.0
numpy==2.0.2
openai==2.9.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.2.3
parsel==1.9.1
Pillow==9.5.0
playwright==1.45.0
playwright-stealth==2.0.0
pycparser==2.23
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.1.0
PyExecJS==1.5.1
PyMySQL==1.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==4.6.0
regex==2025.11.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.37.2
tenacity==8.2.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.29.0
w3lib==2.3.1
wordcloud==1.9.3
yt-dlp
zipp==3.23.0
funasr
modelscope
torchaudio
